lib: Rename functions prefixed with 'geda_line'.
