+++ liblepton/src/line.c
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
-/*! \file geda_line.c
+/*! \file line.c
  *
  *  \brief Low-level mathematical functions for lines
  */
